--- web_ui/packages/config/test/jest.global.mjs
@@ -1,6 +1,0 @@
-// Copyright (C) 2022-2025 Intel Corporation
-// LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
-
-export default async () => {
-    process.env.TZ = 'UTC';
-};
