--- web_ui/packages/ui/eslint.config.mjs
@@ -1,3 +1,0 @@
-import sharedEslintConfig from '@geti/config/lint';
-
-export default sharedEslintConfig;
