--- web_ui/packages/ui/jest.config.mjs
@@ -1,6 +1,0 @@
-// Copyright (C) 2022-2025 Intel Corporation
-// LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
-
-import sharedConfig from '@geti/config/test';
-
-export default sharedConfig;
