+++ web_ui/packages/ui/src/button/button.component.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { forwardRef, Ref } from 'react';
+import { ComponentProps, forwardRef, Ref } from 'react';
 
 import {
     ActionButton as SpectrumActionButton,
@@ -10,6 +10,7 @@ import {
     SpectrumButtonProps,
 } from '@adobe/react-spectrum';
 import { FocusableRef, FocusableRefValue } from '@react-types/shared';
+import { Link } from 'react-router-dom';
 
 type VariantWithoutLegacyButtonVariant = Exclude<SpectrumButtonProps['variant'], 'cta' | 'overBackground'>;
 
@@ -18,12 +19,28 @@ export interface ButtonProps extends Omit<SpectrumButtonProps, 'variant'> {
     variant?: VariantWithoutLegacyButtonVariant;
 }
 
+// https://github.com/adobe/react-spectrum/blob/main/packages/%40react-aria/button/src/useButton.ts#L75
+// This component builds up a link with a fixed `to` prop,
+// this used so that it can be used as `elementTYpe={LinkBuilder(href)}` in a react/spectrum
+// button component, which does not forward an href for custom element types
+type LinkProps = ComponentProps<typeof Link>;
+function LinkBuilder({ href, target, rel }: { href: string; target?: LinkProps['target']; rel: LinkProps['rel'] }) {
+    return forwardRef((props: LinkProps, ref: LinkProps['ref']) => {
+        return <Link {...props} ref={ref} target={target} rel={rel} to={href} />;
+    });
+}
+
 export interface ActionButtonProps extends SpectrumActionButtonProps {
     ref?: Ref<FocusableRefValue<HTMLElement, HTMLButtonElement>>;
 }
 
 export const Button = forwardRef((props: ButtonProps, ref: ButtonProps['ref']) => {
-    return <SpectrumButton {...props} variant={props.variant ?? 'accent'} ref={ref} />;
+    const elementType =
+        props.href === undefined
+            ? props.elementType
+            : LinkBuilder({ href: props.href, target: props.target, rel: props.rel });
+
+    return <SpectrumButton {...props} elementType={elementType} variant={props.variant ?? 'accent'} ref={ref} />;
 });
 
 export const ActionButton = forwardRef((props: ActionButtonProps, ref: ActionButtonProps['ref']) => {
